requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.6.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0